requirements.txt
alembic==1.4.1
aniso8601==8.0.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.2.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coverage==5.0.3
entrypoints==0.3
environs==7.2.0
flake8==3.7.8
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.4.0
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pep8==1.7.1
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
py-healthcheck==1.9.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.3
pytest-cov==2.8.1
pytest-env==0.6.2
pytest-flask==0.15.0
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
sentry-sdk==0.14.1
six==1.12.0
SQLAlchemy==1.3.13
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
workerpool==0.9.4